【frame】框架监控器支持自定义打印信息接口,myclient支持定时发送登陆请求测试

+++ frame/include/FrightenStone/common/net/ProtocolInterface/Defs/NetMsgHeaderFmtType.h
@@ -40,8 +40,8 @@ class BASE_EXPORT NetMsgHeaderFmtType
 {
 public:
     /* 类型定义 */
-    typedef UInt16 PacketLenDataType;       // 包头长度类型 32位64位以及跨平台占用的字节数都是固定的
-    typedef UInt16 CmdDataType;             // 包头协议号类型 32位64位以及跨平台占用的字节数都是固定的
+    typedef UInt32 PacketLenDataType;       // 包头长度类型 32位64位以及跨平台占用的字节数都是固定的
+    typedef UInt32 CmdDataType;             // 包头协议号类型 32位64位以及跨平台占用的字节数都是固定的
 
     // 消息头格式: [4字节长度][4字节cmd]
     static const UInt32 _msgHeaderSize;
